+++ src/hash_map.c
@@ -18,7 +18,6 @@
  */
 
 #include <stdio.h>
-#include <stdint.h>
 #include <stdlib.h>
 #include "slist.h"
 #include "exception.h"
@@ -27,8 +26,8 @@
 #include "rbtree.h"
 #include "hash_map.h"
 
-gds_hash_map_t * gds_hash_map_new(uint32_t size, void *hash_cb, void *cmpkey_cb,
-	void *key_free_cb, void *free_cb)
+gds_hash_map_t * gds_hash_map_new(unsigned long size, void *hash_cb,
+	void *cmpkey_cb, void *key_free_cb, void *free_cb)
 {
 	gds_hash_map_t *h;
 
@@ -55,14 +54,14 @@ gds_hash_map_t * gds_hash_map_new(uint32_t size, void *hash_cb, void *cmpkey_cb,
 	return h;
 }
 
-uint32_t gds_hash_map_hash(gds_hash_map_t *h, const void *key)
+unsigned long gds_hash_map_hash(gds_hash_map_t *h, const void *key)
 {
 	return h->hash_cb(key, h->size) % h->size;
 }
 
 int gds_hash_map_set(gds_hash_map_t *h, void *key, void *data)
 {
-	uint32_t hash;
+	unsigned long hash;
 	int rc;
 
 	GDS_CHECK_ARG_NOT_NULL(h);
@@ -76,7 +75,7 @@ int gds_hash_map_set(gds_hash_map_t *h, void *key, void *data)
 
 void * gds_hash_map_get(gds_hash_map_t *h, const void *key)
 {
-	uint32_t hash;
+	unsigned long hash;
 	void *data;
 
 	GDS_CHECK_ARG_NOT_NULL(h);
@@ -88,7 +87,7 @@ void * gds_hash_map_get(gds_hash_map_t *h, const void *key)
 
 int gds_hash_map_unset(gds_hash_map_t *h, const void *key)
 {
-	uint32_t hash;
+	unsigned long hash;
 	int rv;
 
 	GDS_CHECK_ARG_NOT_NULL(h);
@@ -102,14 +101,14 @@ int gds_hash_map_unset(gds_hash_map_t *h, const void *key)
 
 typedef struct {
 	gds_hash_map_t *hash;
-	uint32_t i;
+	unsigned long i;
 	gds_iterator_t *rbtree_it;
 } gds_hash_map_iterator_data_t;
 
 void gds_hash_map_iterator_find_next_rbtree(gds_hash_map_iterator_data_t *data)
 {
 	gds_hash_map_t *hash = data->hash;
-	uint32_t i = data->i;
+	unsigned long i = data->i;
 
 	while (i < hash->size && hash->map[i] == NULL)
 		i++;
@@ -203,7 +202,7 @@ gds_slist_t * gds_hash_map_keys(gds_hash_map_t *h)
 {
 	gds_slist_t *l = gds_slist_new();
 	gds_slist_t *list;
-	uint32_t i;
+	unsigned long i;
 
 	GDS_CHECK_ARG_NOT_NULL(h);
 
@@ -222,7 +221,7 @@ gds_slist_t * gds_hash_map_values(gds_hash_map_t *h)
 {
 	gds_slist_t *l = gds_slist_new();
 	gds_slist_t *list;
-	uint32_t i;
+	unsigned long i;
 
 	GDS_CHECK_ARG_NOT_NULL(h);
 
@@ -237,11 +236,12 @@ gds_slist_t * gds_hash_map_values(gds_hash_map_t *h)
 	return l;
 }
 
-gds_rbtree_node_t ** gds_hash_map_build_map(gds_hash_map_t *h, uint32_t size)
+gds_rbtree_node_t ** gds_hash_map_build_map(gds_hash_map_t *h,
+	unsigned long size)
 {
 	gds_rbtree_node_t **map;
 	gds_iterator_t *it;
-	uint32_t hash;
+	unsigned long hash;
 	void *k, *v;
 
 	map = calloc(size, sizeof(gds_rbtree_node_t *));
@@ -261,7 +261,7 @@ gds_rbtree_node_t ** gds_hash_map_build_map(gds_hash_map_t *h, uint32_t size)
 	return map;
 }
 
-void gds_hash_map_change_size(gds_hash_map_t *h, uint32_t new_size)
+void gds_hash_map_change_size(gds_hash_map_t *h, unsigned long new_size)
 {
 	gds_rbtree_node_t **map;
 
@@ -269,7 +269,7 @@ void gds_hash_map_change_size(gds_hash_map_t *h, uint32_t new_size)
 	GDS_CHECK_ARG_NOT_ZERO(new_size);
 
 	map = gds_hash_map_build_map(h, new_size);
-	for (uint32_t i = 0; i < h->size; i++) {
+	for (unsigned long i = 0; i < h->size; i++) {
 		gds_rbtree_free(h->map[i], NULL, NULL);
 	}
 	free(h->map);
@@ -279,7 +279,7 @@ void gds_hash_map_change_size(gds_hash_map_t *h, uint32_t new_size)
 
 void gds_hash_map_free(gds_hash_map_t *h)
 {
-	uint32_t i;
+	unsigned long i;
 	if (h != NULL) {
 		for (i=0; i<h->size; i++) {
 			gds_rbtree_free(h->map[i], h->key_free_cb, h->free_cb);
